administration: add ui schema
* form: order fields according to config
* fix file structure
* add formatter on the details page
