+++ invenio_administration/assets/semantic-ui/js/invenio_administration/src/actions/ActionsDropdown.js
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 import { Dropdown, Button } from "semantic-ui-react";
 import { ResourceActions } from "./ResourceActions";
 import isEmpty from "lodash/isEmpty";
-import { i18next } from "../../../translations/invenio_administration/i18next";
-import { DeleteModalTrigger } from "../../actions/DeleteModalTrigger";
+import { i18next } from "@translations/invenio_administration/i18next";
+import { DeleteModalTrigger } from "./DeleteModalTrigger";
 
 export class ActionsDropdown extends Component {
   render() {
