+++ invenio_administration/assets/semantic-ui/js/invenio_administration/src/actions/DeleteModalTrigger.js
