+++ invenio_administration/assets/semantic-ui/js/invenio_administration/src/details/ActionModal.js
