+++ invenio_administration/assets/semantic-ui/js/invenio_administration/src/ui_messages/messages.js
