+++ invenio_administration/templates/semantic-ui/invenio_administration/details.html
@@ -11,8 +11,7 @@
 
 {% block page_title %}
   {{ go_back() }}
-  <h1>{{ title or name }}</h1>
-  <div class="ui divider"></div>
+
 {% endblock page_title %}
 
 {% block admin_page_content %}
