+++ invenio_administration/templates/semantic-ui/invenio_administration/macros.html
@@ -1,15 +1,15 @@
 {#
     Copyright (C) 2022 CERN.
-  
+
     Invenio Administration is free software; you can redistribute it and/or modify it
     under the terms of the MIT License; see LICENSE file for more details.
   #}
-  
+
 
 {% macro go_back() %}
     {% if request.referrer %}
     <div class="mb-10">
-    <a href="{{ request.referrer }}" class="ui small button labeled icon rel-mr-1">
+    <a href="{{ request.referrer }}" class="ui mini button labeled icon rel-mr-1">
         <i class="ui icon left arrow" aria-hidden="true"></i>
         {{ _('Back')}}
     </a>
