+++ invenio_administration/views/base.py
@@ -22,7 +22,10 @@ from invenio_administration.errors import (
     MissingExtensionName,
     MissingResourceConfiguration,
 )
-from invenio_administration.marshmallow_utils import jsonify_schema
+from invenio_administration.marshmallow_utils import (
+    jsonify_schema,
+    set_ui_field_property,
+)
 
 
 class AdminView(MethodView):
@@ -224,11 +227,23 @@ class AdminResourceBaseView(AdminView):
         """Returns administration UI list view endpoint."""
         if self.list_view_name:
             return url_for(f"administration.{self.list_view_name}")
+        if isinstance(self, AdminResourceListView):
+            return url_for(f"administration.{self.name}")
 
     def get_create_view_endpoint(self):
         """Returns administration UI list view endpoint."""
         if self.create_view_name:
             return url_for(f"administration.{self.create_view_name}")
+
+    def set_ui_schema_params(self, jsonschema, config):
+        """Add UI properties to resource schema."""
+        if not config:
+            return jsonschema
+
+        ui_schema = {}
+        for field_path, field_properties in config.items():
+            set_ui_field_property(jsonschema, field_path, field_properties, ui_schema)
+        return ui_schema
 
 
 class AdminResourceDetailView(AdminResourceBaseView):
@@ -249,11 +264,12 @@ class AdminResourceDetailView(AdminResourceBaseView):
         schema = self.get_service_schema()
         serialized_schema = self._schema_to_json(schema)
         fields = self.item_field_list
+        schema = self.set_ui_schema_params(serialized_schema, self.item_field_list)
         # TODO context processor?
         return self.render(
-            ** {
+            **{
                 "name": name,
-                "resource_schema": serialized_schema,
+                "resource_schema": schema,
                 "fields": fields,
                 "exclude_fields": self.item_field_exclude_list,
                 "pid": pid_value,
@@ -283,11 +299,11 @@ class AdminFormView(AdminResourceBaseView):
         schema = self.get_service_schema()
         serialized_schema = self._schema_to_json(schema)
         form_fields = self.form_fields
-
+        ui_schema = self.set_ui_schema_params(serialized_schema, self.form_fields)
         # TODO context processor?
         return self.render(
             **{
-                "resource_schema": serialized_schema,
+                "resource_schema": ui_schema,
                 "form_fields": form_fields,
                 "pid": pid_value,
                 "api_endpoint": self.get_api_endpoint(),
@@ -333,7 +349,7 @@ class AdminResourceListView(AdminResourceBaseView):
     api_endpoint = None
     title = None
 
-    search_request_headers = {"Accept": "application/vnd.inveniordm.v1+json"}
+    search_request_headers = {"Accept": "application/json"}
 
     def get_search_request_headers(self):
         """Get search request headers."""
